requirements/common.txt
# requirements/common.txt
Django==1.4.5
cssutils==0.9.10b1
django-tastypie==0.9.12
mimeparse==0.1.3
python-dateutil==2.1
six==1.2.0